Module parachain_gen::parachain_runtime

Modules

aura

aura_ext

authorship

balances

beefy

beefy_light_client

beefy_mmr

bridge_data_signer

collator_selection

cumulus_xcm

dmp_queue

leaf_provider

mmr

multisig_verifier

parachain_info

parachain_system

polkadot_xcm

runtime_types

session

substrate_bridge_inbound_channel

substrate_bridge_outbound_channel

substrate_dispatch

sudo

system

timestamp

transaction_payment

x_tokens

xcm_app

xcm_app_sudo_wrapper

xcmp_queue

Structs

ConstantsApi

StorageApi

TransactionApi

Enums

Event

Statics

PALLETS

Functions

constants

storage

tx

validate_codegen

check whether the Client you are using is aligned with the statically generated codegen.

Type Definitions

DispatchError

The default error type returned when there is a runtime issue, exposed here for ease of use.