Module parachain_gen::parachain_runtime::polkadot_xcm::events

Structs

AssetsClaimed

Some assets have been claimed from an asset trap

AssetsTrapped

Some assets have been placed in an asset trap.

Attempted

Execution of an XCM message was attempted.

FeesPaid

Fees were paid from a location for an operation (often for using SendXcm).

InvalidQuerier

Expected query response has been received but the querier location of the response does not match the expected. The query remains registered for a later, valid, response to be received and acted upon.

InvalidQuerierVersion

Expected query response has been received but the expected querier location placed in storage by this runtime previously cannot be decoded. The query remains registered.

InvalidResponder

Expected query response has been received but the origin location of the response does not match that expected. The query remains registered for a later, valid, response to be received and acted upon.

InvalidResponderVersion

Expected query response has been received but the expected origin location placed in storage by this runtime previously cannot be decoded. The query remains registered.

Notified

Query response has been received and query is removed. The registered notification has been dispatched and executed successfully.

NotifyDecodeFailed

Query response has been received and query is removed. The dispatch was unable to be decoded into a Call; this might be due to dispatch function having a signature which is not (origin, QueryId, Response).

NotifyDispatchError

Query response has been received and query is removed. There was a general error with dispatching the notification call.

NotifyOverweight

Query response has been received and query is removed. The registered notification could not be dispatched because the dispatch weight is greater than the maximum weight originally budgeted by this runtime for the query result.

NotifyTargetMigrationFail

A given location which had a version change subscription was dropped owing to an error migrating the location to our new XCM format.

NotifyTargetSendFail

A given location which had a version change subscription was dropped owing to an error sending the notification to it.

ResponseReady

Query response has been received and is ready for taking with take_response. There is no registered notification call.

ResponseTaken

Received query response has been read and removed.

Sent

A XCM message was sent.

SupportedVersionChanged

The supported version of a location has been changed. This might be through an automatic notification or a manual intervention.

UnexpectedResponse

Query response received which does not match a registered query. This may be because a matching query was never registered, it may be because it is a duplicate response, or because the query timed out.

VersionChangeNotified

An XCM version change notification message has been attempted to be sent.

VersionNotifyRequested

We have requested that a remote chain sends us XCM version change notifications.

VersionNotifyStarted

A remote has requested XCM version change notification from us and we have honored it. A version information message is sent to them and its cost is included.

VersionNotifyUnrequested

We have requested that a remote chain stops sending us XCM version change notifications.